[#188] refactor based on Rubocop warnings
Mount `.rubocop.yml` in dev, for convenience when reducing warnings or
improving exclusions.

Trailing commas might be contentious, since Ruby supports both. For
arrays, this doesn't seem so bad, and is in fact often useful. But for
method calls, this can be (in my opinion) odd, depending on which
function of indentation for the function call is used. Here I've opted
to rewrite all the trailing commas, but if this is problematic, we can
switch to mandatory trailing commas for all Register repos instead.

`Parsers::JsonReader` `foreach` signature is changed, since `header` was
unimplemented, and unused everywhere.

There are some functions with high complexity warnings. On inspection, I
agree that they appear to be rather complex. However, I'm not rewriting
them within this patch, since I'm looking for fairly-quick improvements.

There are significant warnings coming from `spec/` tests. It would seem
that extensive changes to the test structure would be required in order
for those to pass. Since I'm unconvinced by the utility of doing this at
the present time, I instead continue to ignore most of those warnings
for tests only. However, for ones which occur only in specific places, I
prefer to exclude them locally, instead.