[#171] add Utils::Object compact_deep
This compacts an object, optionally pruning empty values such as empty
hashes, empty arrays, or empty strings. This is useful when dealing
with source data containing JSON null values, since the absence of a key
is not the same as the presence of a key with a null value.

This was highlighted during work on importing AM data via static BODS.