fix Elasticsearch::Query search_scroll for changing scroll_id
The example code for Ruby in the Elasticsearch documentation says:

    # Search request with a scroll argument:
    response = client.search(index: 'books', scroll: '10m')

    # Save the scroll id:
    scroll_id = response['_scroll_id']

    # Print out the names of all the books we find:
    while response['hits']['hits'].size.positive?
      response = client.scroll(scroll: '5m', body: { scroll_id: scroll_id })
      puts(response['hits']['hits'].map { |r| r['_source']['name'] })
    end

https://www.elastic.co/guide/en/elasticsearch/client/ruby-api/current/examples.html#ex-scroll

This fetches `scroll_id`, and uses it for all subsequent requests.

And yet, the main API Elasticsearch documentation says:

    The initial search request and each subsequent scroll request each return a _scroll_id. While the _scroll_id may change between requests, it doesn’t always change — in any case, only the most recently received _scroll_id should be used.

https://www.elastic.co/guide/en/elasticsearch/reference/current/paginate-search-results.html#scroll-search-results

This says that `scroll_id` might in change change mid-way through the
scroll.

Fix to update `scroll_id` to be the latest response's, not the one found
in the beginning.

This bug has been present since 2023-11-17 in
e9b76c1caa9efb0b183dd375f6dfa4595c5b3120 , which was an extraction of the
same code from Sources OC. That bug was present since 2023-08-30 in
b665f862e88cbb09d4942b7d83b68d07b47784bb .