[#249] unset TTL in StreamClientKinesis
I'd rather Transformer PSC errored in the case of an old or invalid
pointer, similar to Ingester PSC, without the TTL silently vanishing.

Additionally, setting the TTL depends on the data retention period of
the Kinesis stream; although it is currently set to 24h, it would be
nice to be able to extend this without requiring code changes.